Added logic to enforce Twig and TwigComponent extensions in `IbexaTwigComponentsBundle` and updated related tests
