+++ src/bundle/IbexaTwigComponentsBundle.php
@@ -9,6 +9,7 @@ declare(strict_types=1);
 namespace Ibexa\Bundle\TwigComponents;
 
 use Ibexa\Bundle\TwigComponents\DependencyInjection\Compiler\ComponentPass;
+use LogicException;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -16,6 +17,14 @@ final class IbexaTwigComponentsBundle extends Bundle
 {
     public function build(ContainerBuilder $container): void
     {
+        if (!$container->hasExtension('twig')) {
+            throw new LogicException('IbexaTwigComponentsBundle requires TwigBundle. Please enable it in your bundles.php file.');
+        }
+
+        if (!$container->hasExtension('twig_component')) {
+            throw new LogicException('IbexaTwigComponentsBundle requires TwigComponentBundle (symfony/ux-twig-component). Please enable it in your bundles.php file.');
+        }
+
         $container->addCompilerPass(new ComponentPass());
     }
 }
